pyproject.toml
[project]
name = "tortoise-auth"
version = "0.1.0"
description = "Async authentication and user management for Tortoise ORM. Framework-agnostic, extensible, secure by default."
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Elie" }]
keywords = ["tortoise-orm", "authentication", "auth", "jwt", "async", "user-management", "totp", "2fa", "api-keys"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "tortoise-orm>=0.21",
    "pwdlib[argon2,bcrypt]>=0.2",
    "PyJWT>=2.8",
    "pyotp>=2.9",
    "qrcode[pil]>=7.4",
]

[project.urls]
Homepage = "https://github.com/elie/tortoise-auth"
Documentation = "https://elie.github.io/tortoise-auth"
Repository = "https://github.com/elie/tortoise-auth"
Issues = "https://github.com/elie/tortoise-auth/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tortoise_auth"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "aiosqlite>=0.20",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["tortoise_auth"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tortoise.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pwdlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyotp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qrcode.*"
ignore_missing_imports = true
